Fix AMD gpu users booting into black screen
Without this patch, AMD gpu users boot into a black screen on the first boot after installing Stephenson's Rocket. It checks if an AMD gpu is in the system and if no NVidia gpu is in the system, if both are true radeon is blacklisted. This prevents the radeon driver from locking up the system.