Created addtopool.sh
This script can move a bunch of deb files from any directory into the pool in the correct directories. To find the correct directories, it currently looks at the package lists in the alchemist, alchemist beta and wheezy repos. If I end up needing it, I will add wheezy backports to that later. The -i option makes you able to add newer version of packages than available in any of the repos, though. The -n option forces the script to created directories based on package name rather than anything existing in the repos, preferably use the -i option in combination with this one. I hope this helps, check the usage for more info.