Many changes to gen.sh
Updated to SteamOS 153.
Now moves packages which are in both the iso and the pool out of the pool.
Obsoletereport was removed, since download-updates.sh replaces both it and the archive-mirror.sh script.
Fixed a bug with gen.sh removing the wrong versions of packages from the buildroot when the a + symbol is found in the latest version number.
Made isohdpfx.bin load correctly on Ubuntu 14.10+. It can now also be loaded from the current working directory if not found anywhere else.
Switched some things from createbuildroot to the createiso function.
Removed realpath from the dependency list, since it's no longer used.

I think that's about it. The firmware.txt file still needs to be updated, though.